+++ resources/js/areaChart.js
@@ -89,6 +89,9 @@ const areaChart = () => {
                         bottom: 0,
                     }
                 },
+
+                theme: component.get('areaChartModel.theme') || {},
+
             };
 
             this.chart = new ApexCharts(this.$refs.container, options);