+++ src/Models/PieChartModel.php
@@ -9,13 +9,13 @@ namespace Asantibanez\LivewireCharts\Models;
  */
 class PieChartModel extends BaseChartModel
 {
-    public $onSliceClickEventName;
+    private $onSliceClickEventName;
 
-    public $opacity;
+    private $opacity;
 
-    public $type;
+    private $type;
 
-    public $data;
+    private $data;
 
     public function __construct()
     {
@@ -42,6 +42,16 @@ class PieChartModel extends BaseChartModel
         $this->type = $type;
 
         return $this;
+    }
+
+    public function asPie()
+    {
+        return $this->setType('pie');
+    }
+
+    public function asDonut()
+    {
+        return $this->setType('donut');
     }
 
     public function withOnSliceClickEvent($onSliceClickEventName)